Restore Kokkos_InitializationSettings.hpp
